feat: implement app store architecture and web storefront
- Defined modular Clean Architecture in ARCHITECTURE.md.
- Built Android MVP in aurora-next/ with Compose and Hilt.
- Created Next.js 16 web storefront with React 19 and Tailwind 4.
- Optimized for mobile-first UI and verified with Playwright.

+++ app/src/main/java/com/aurora/store/data/model/SessionInfo.kt
@@ -3,6 +3,6 @@ package com.aurora.store.data.model
 data class SessionInfo(
     val sessionId: Int,
     val packageName: String,
-    val versionCode: Long,
+    val versionCode: Int,
     val displayName: String = String()
 )
